requirements.txt
streamlit
chromadb
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-huggingface
sentence-transformers
pypdf
python-dotenv
# Choose an LLM provider (here: Groq). You can swap later.
langchain-groq
